+++ ZComponents/ZCrossProductTerm.h
@@ -16,7 +16,7 @@ public:
 
    // FUNCS
     bool zp_calcValue(const ZAbstractSpectrum*, qreal&) override;
-    bool zp_calcTermVariablePart(const ZAbstractSpectrum*, qint64 &value) override; // w/o factor
+    bool zp_calcTermVariablePart(const ZAbstractSpectrum*, qreal &value) override; // w/o factor
     bool zp_termBelongsToWindow(const ZCalibrationWindow*) const override;
 
    QString zp_termName() const override;
